Move ClientCache to source-shared class
